packages/agent-core/requirements.txt
# Core Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.1.0

# AI Integration
google-genai>=1.0.0

# HTTP Client
httpx>=0.27.0

# Environment
python-dotenv>=1.0.0

# Docker
docker>=7.0.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Code Quality (add these for later)
pylint>=3.0.3
mypy>=1.8.0
black>=24.1.1